# contracts/openapi/ledger.v1.yaml
openapi: 3.1.0
info:
  title: LedgerCore Ledger API
  version: 1.0.0
  description: |
    Double-entry ledger API. All monetary amounts are integers in minor units,
    string-encoded to avoid int64 precision loss in JSON, together with an
    asset code. Every transaction must balance debits and credits per asset.

    Multi-tenancy: the tenant is derived from the JWT; resources are always
    scoped to the authenticated tenant.
  contact:
    name: LedgerCore Platform Team
    url: https://github.com/ledgercore/ledgercore
servers:
  - url: http://localhost:8081
    description: Local development (direct)
  - url: http://localhost:8080
    description: Local development (via gateway; paths are routed unchanged)
security:
  - bearerAuth: []
...
paths:
  /v1/ledgers:
    post:
      operationId: createLedger
      tags: [Ledgers]
      summary: Create a ledger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerCreate'
            examples:
              main:
                summary: Main USD ledger
                value:
                  name: main
                  description: Primary operating ledger
                  metadata:
                    region: us
      responses:
        '201':
          description: Ledger created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ledger'
              examples:
                main:
                  value:
                    id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                    name: main
                    description: Primary operating ledger
                    environment: sandbox
                    metadata:
                      region: us
                    created_at: '2026-07-24T12:00:00Z'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      operationId: listLedgers
      tags: [Ledgers]
      summary: List ledgers
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page of ledgers.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ledger'
                  next_cursor:
                    type: [string, 'null']
                    description: Opaque cursor for the next page; null when exhausted.
              examples:
                page:
                  value:
                    data:
                      - id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                        name: main
                        description: Primary operating ledger
                        environment: sandbox
                        metadata: {}
                        created_at: '2026-07-24T12:00:00Z'
                    next_cursor: null
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/ledgers/{id}:
    get:
      operationId: getLedger
      tags: [Ledgers]
      summary: Get a ledger
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The ledger.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ledger'
              examples:
                main:
                  value:
                    id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                    name: main
                    description: Primary operating ledger
                    environment: sandbox
                    metadata: {}
                    created_at: '2026-07-24T12:00:00Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/accounts:
    post:
      operationId: createAccount
      tags: [Accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
            examples:
              customerWallet:
                summary: Customer wallet account
                value:
                  ledger_id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                  name: customer:cust_42:wallet
                  type: liability
                  normal_balance: CREDIT
                  metadata:
                    customer_id: cust_42
      responses:
        '201':
          description: Account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
              examples:
                customerWallet:
                  value:
                    id: 2b1c7d3f-0a2b-7c4d-9e50-6f8a0b1c2d3e
                    ledger_id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                    name: customer:cust_42:wallet
                    type: liability
                    normal_balance: CREDIT
                    status: active
                    metadata:
                      customer_id: cust_42
                    created_at: '2026-07-24T12:01:00Z'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      operationId: listAccounts
      tags: [Accounts]
      summary: List accounts
      parameters:
        - name: ledger_id
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AccountType'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page of accounts.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
                  next_cursor:
                    type: [string, 'null']
              examples:
                page:
                  value:
                    data:
                      - id: 2b1c7d3f-0a2b-7c4d-9e50-6f8a0b1c2d3e
                        ledger_id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                        name: customer:cust_42:wallet
                        type: liability
                        normal_balance: CREDIT
                        status: active
                        metadata: {}
                        created_at: '2026-07-24T12:01:00Z'
                    next_cursor: null
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/accounts/{id}:
    get:
      operationId: getAccount
      tags: [Accounts]
      summary: Get an account
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
              examples:
                customerWallet:
                  value:
                    id: 2b1c7d3f-0a2b-7c4d-9e50-6f8a0b1c2d3e
                    ledger_id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                    name: customer:cust_42:wallet
                    type: liability
                    normal_balance: CREDIT
                    status: active
                    metadata: {}
                    created_at: '2026-07-24T12:01:00Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/transactions/{id}:
    get:
      operationId: getTransaction
      tags: [Transactions]
      summary: Get a transaction
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
              examples:
                usdDeposit:
                  value:
                    id: 3c2d8e0f-1a2b-7c4d-9e5f-7a8b9c0d1e2f
                    ledger_id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                    idempotency_key: dep-2026-0001
                    description: Customer USD deposit
                    status: posted
                    postings:
                      - account_id: 5e4f0a2b-3c4d-7e6f-b072-9a0b1c2d3e4f
                        direction: DEBIT
                        amount: { asset: USD, amount: '10000' }
                      - account_id: 2b1c7d3f-0a2b-7c4d-9e50-6f8a0b1c2d3e
                        direction: CREDIT
                        amount: { asset: USD, amount: '9700' }
                      - account_id: 6f5a1b3c-4d5e-7f80-c193-0b1c2d3e4f5a
                        direction: CREDIT
                        amount: { asset: USD, amount: '300' }
                    metadata: {}
                    reversed_by: null
                    reverses: null
                    effective_at: '2026-07-24T12:05:00Z'
                    created_at: '2026-07-24T12:05:00Z'
                    posted_at: '2026-07-24T12:05:00Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/transactions/{id}/post:
    post:
      operationId: postTransaction
      tags: [Transactions]
      summary: Post a draft transaction
      description: Transitions a draft transaction to posted. Idempotent for already-posted transactions.
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The transaction, now posted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
              examples:
                posted:
                  value:
                    id: 3c2d8e0f-1a2b-7c4d-9e5f-7a8b9c0d1e2f
                    ledger_id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                    idempotency_key: dep-2026-0001
                    description: Customer USD deposit
                    status: posted
                    postings: []
                    metadata: {}
                    reversed_by: null
                    reverses: null
                    effective_at: '2026-07-24T12:05:00Z'
                    created_at: '2026-07-24T12:05:00Z'
                    posted_at: '2026-07-24T12:06:00Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The transaction is reversed and cannot be posted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                reversed:
                  value:
                    error:
                      code: conflict
                      message: transaction is reversed and cannot be posted
...
  /v1/holds/{id}:
    get:
      operationId: getHold
      tags: [Holds]
      summary: Get a hold
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The hold.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hold'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/holds/{id}/release:
    post:
      operationId: releaseHold
      tags: [Holds]
      summary: Release a hold
      description: Frees the reserved funds without moving money.
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Hold released.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hold'
              examples:
                released:
                  value:
                    id: 8b7c3d5e-6f7a-7b8c-e3b5-2d3e4f5a6b7c
                    ledger_id: 1f0a6a2e-9c1b-7f2e-8a11-3c5d7e9f0b21
                    account_id: 2b1c7d3f-0a2b-7c4d-9e50-6f8a0b1c2d3e
                    idempotency_key: auth-2026-0042
                    amount: { asset: USD, amount: '2500' }
                    status: released
                    transaction_id: null
                    expires_at: '2026-07-31T00:00:00Z'
                    metadata: {}
                    created_at: '2026-07-24T13:00:00Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The hold is not active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notActive:
                  value:
                    error:
                      code: conflict
                      message: hold is already captured
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: EdDSA JWT issued by the identity service (POST /v1/auth/token).
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Page size; values above 200 are clamped to 200.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    BadRequest:
      description: Invalid input (`validation_failed`) or malformed pagination cursor (`invalid_cursor`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            badRequest:
              value:
                error:
                  code: validation_failed
                  message: 'unknown field "amonut"'
                  request_id: 9f1b2c3d4e5f60718293a4b5c6d7e8f9
...
  schemas:
...
    LedgerCreate:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1024
        metadata:
          $ref: '#/components/schemas/Metadata'
    Ledger:
      type: object
      required: [id, name, environment, metadata, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        environment:
          type: string
          enum: [sandbox, live]
          description: Taken from the `env` claim of the creating token; not client-settable.
        metadata:
          $ref: '#/components/schemas/Metadata'
        created_at:
          type: string
          format: date-time
...